set up delete comment only if userAuth and commentatorInfo userId is === req.user.id which comes from authMiddleware

+++ src/services/comments-service.ts
@@ -1,10 +1,16 @@
-import { CommentInputModel } from "../models";
+import { CommentInputModel, UserViewModel } from "../models";
 import { commentsRepository } from "../repositories";
 
 export const commentsService = {
-  async removeComment(commentId: string) {
-    const foundComment = await commentsRepository.removeComment(commentId);
-    return foundComment;
+  async removeComment(commentId: string, user: UserViewModel) {
+    const foundComment = await commentsRepository.getByIdComment(commentId);
+
+    if (foundComment && user.id !== foundComment.commentatorInfo.userId) {
+      return 403;
+    }
+
+    const commentToRemove = await commentsRepository.removeComment(commentId);
+    return commentToRemove;
   },
 
   async updateComment(data: CommentInputModel, commentId: string) {
